In daemonshepherd, don't remove socket if it was TCP on close.

+++ lib/panopticon/daemonshepherd/control_socket.py
@@ -39,7 +39,8 @@ class ControlSocket:
     if self.socket is not None:
       self.socket.close()
       self.socket = None
-      os.remove(self.path)
+      if self.path is not None:
+        os.remove(self.path)
 
 #-----------------------------------------------------------------------------
